refactor: move readline stuff into its own file
This is in direct preparation for an automation mode of nix repl.
